Create separate search module.
